// frontend/src/app/api/model/featureAssignmentDto.ts
/**
 * Api
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface FeatureAssignmentDto { 
    stageId?: number;
    stageName?: string | null;
    userId?: number | null;
    userName?: string | null;
    estimation?: number | null;
    estimationUnitName?: string | null;
    startedAt?: string | null;
    completedAt?: string | null;
}