// frontend/src/app/api/model/featureStageLogOutput.ts
/**
 * Api
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface FeatureStageLogOutput { 
    id?: number;
    action?: string | null;
    comment?: string | null;
    timestamp?: string;
    userId?: number;
    userName?: string | null;
    stageName?: string | null;
}